+++ src/services/LeadConfigService.js
@@ -2,6 +2,7 @@ const LeadStatus = require("../models/LeadStatus");
 const LeadStatusGroup = require("../models/LeadStatusGroup");
 const { generateMonotonicId, ID_PREFIXES } = require("../utils/id");
 const { createHttpError } = require("../utils/http");
+const { isSystemEntity } = require("../constants/systemFunnel");
 
 class LeadConfigService {
   async getStatuses() {
@@ -18,6 +19,7 @@ class LeadConfigService {
   }
 
   async updateStatus(id, data) {
+    if (isSystemEntity(id)) throw createHttpError(400, "Không thể sửa trạng thái hệ thống.");
     const updatedStatus = await LeadStatus.findOneAndUpdate({ id }, data, { new: true });
     if (!updatedStatus) {
       throw createHttpError(404, "Không tìm thấy trạng thái");
@@ -26,6 +28,8 @@ class LeadConfigService {
   }
 
   async deleteStatus(id) {
+    if (isSystemEntity(id)) throw createHttpError(400, "Không thể xoá trạng thái hệ thống.");
+
     const usedInGroup = await LeadStatusGroup.findOne({ statusIds: id });
     if (usedInGroup) {
       throw createHttpError(400, "Không thể xóa trạng thái đang được sử dụng trong nhóm.");
@@ -52,6 +56,7 @@ class LeadConfigService {
   }
 
   async updateGroup(id, data) {
+    if (isSystemEntity(id)) throw createHttpError(400, "Không thể sửa nhóm trạng thái hệ thống.");
     const updatedGroup = await LeadStatusGroup.findOneAndUpdate({ id }, data, { new: true });
     if (!updatedGroup) {
       throw createHttpError(404, "Không tìm thấy nhóm trạng thái");
@@ -60,6 +65,8 @@ class LeadConfigService {
   }
 
   async deleteGroup(id) {
+    if (isSystemEntity(id)) throw createHttpError(400, "Không thể xoá nhóm trạng thái hệ thống.");
+
     const deletedGroup = await LeadStatusGroup.findOneAndDelete({ id });
     if (!deletedGroup) {
       throw createHttpError(404, "Không tìm thấy nhóm trạng thái");
